requirements.txt
pyspark==3.5.5
kafka-python==2.0.2
confluent-kafka==2.3.0
tensorflow==2.13.0
keras==2.13.1
nltk==3.8.1
joblib==1.3.2
cassandra-driver==3.28.0